windows: IsConsole(): fix deprecation comment
This function was deprecated in 57a2131dd69588daf506fde11b5c51ac6e271025,
but the comment was not formatted correctly, causing it to not be documented
as deprecated on pkg.go.dev.

Also changing this back to an actual function wrapper, so that it doesn't
get documented in the "Variables" section.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>